Moved the test to com.jcabi.email.EnvelopeTest.handlesUnicodeCorrectly.
